test: migrate http test
